Revert "Change BIGINT to INTEGER"
This reverts commit c1225f1a9e1e62baf8220e6af9f07d119a216c3a.

+++ src/data/migrations/20190226111416-add-system-metrics-to-fog.js
@@ -3,10 +3,10 @@
 module.exports = {
   up: (queryInterface, Sequelize) => {
     return queryInterface.addColumn('Fogs',
-      'system-available-memory', Sequelize.INTEGER
+      'system-available-memory', Sequelize.BIGINT
     ).then(() => {
       return queryInterface.addColumn('Fogs',
-        'system-available-disk', Sequelize.INTEGER
+        'system-available-disk', Sequelize.BIGINT
       )
     }).then(() => {
       return queryInterface.addColumn('Fogs',
